0.8.57.0 Wanamingo data pages added and bugs fixed

+++ Hearts Of Greed/Motor2D/Enemy.h
@@ -72,7 +72,7 @@ public:
 
 	int GetHP();
 	int GetAD();
-	int GetAS();
+	float GetAS();
 	int GetVision();
 	int GetRecov();
